firmware: move hackrf_core drivers & configurations to drivers.h|c
