+++ firmware/common/clock_gen.c
@@ -23,7 +23,7 @@
 
 #include <stdint.h>
 
-#include "hackrf_core.h"
+#include "drivers.h"
 #include "hackrf_ui.h"
 #include "i2c_bus.h"
 #include "platform_detect.h"
