+++ firmware/common/fpga_selftest.c
@@ -25,9 +25,9 @@
 
 #include "clock_gen.h"
 #include "delay.h"
+#include "drivers.h"
 #include "fixed_point.h"
 #include "fpga.h"
-#include "hackrf_core.h"
 #include "ice40_spi.h"
 #include "m0_state.h"
 #include "max283x.h"
