+++ firmware/common/hackrf_core.h
@@ -27,50 +27,6 @@
 extern "C" {
 #endif
 
-#include "cpld_jtag.h"
-#include "i2c_bus.h"
-#include "i2c_lpc.h"
-#include "max283x.h"
-#include "max5864.h"
-#include "mixer.h"
-#include "platform_detect.h" // IWYU pragma: keep
-#include "radio.h"
-#include "rf_path.h"
-#include "sgpio.h"
-#include "si5351c.h"
-#include "spi_ssp.h"
-#include "w25q80bv.h"
-#ifdef IS_PRALINE
-	#include "fpga.h"
-	#include "ice40_spi.h"
-#endif
-
-/* TODO: Hide these configurations */
-extern const i2c_lpc_config_t i2c_config_si5351c_fast_clock;
-extern si5351c_driver_t clock_gen;
-extern ssp_config_t ssp_config_w25q80bv;
-
-extern max283x_driver_t max283x;
-#ifdef IS_PRALINE
-extern ice40_spi_driver_t ice40;
-extern fpga_driver_t fpga;
-#endif
-extern max283x_driver_t max283x;
-extern max5864_driver_t max5864;
-extern mixer_driver_t mixer;
-extern w25q80bv_driver_t spi_flash;
-extern sgpio_config_t sgpio_config;
-extern radio_t radio;
-extern rf_path_t rf_path;
-extern jtag_t jtag_cpld;
-extern i2c_bus_t i2c0;
-
-void ssp1_set_mode_max283x(void);
-void ssp1_set_mode_max5864(void);
-#ifdef IS_PRALINE
-void ssp1_set_mode_ice40(void);
-#endif
-
 void pin_shutdown(void);
 void pin_setup(void);
