+++ firmware/common/operacake.c
@@ -26,9 +26,9 @@
 
 #include <libopencm3/lpc43xx/scu.h>
 
+#include "drivers.h"
 #include "gpio.h"
 #include "gpio_lpc.h"
-#include "hackrf_core.h"
 #include "i2c_bus.h"
 #include "operacake_sctimer.h"
 #include "platform_scu.h"
