+++ firmware/common/rf_path.c
@@ -23,7 +23,7 @@
 
 #include "rf_path.h"
 
-#include "hackrf_core.h"
+#include "drivers.h"
 #include "max283x.h"
 #include "max5864.h"
 #include "mixer.h"
