+++ firmware/common/rom_iap.c
@@ -22,7 +22,7 @@
 
 #include <stdint.h>
 
-#include "hackrf_core.h"
+#include "drivers.h"
 #include "rom_iap.h"
 #include "spi_bus.h"
 #include "w25q80bv.h"
