+++ firmware/hackrf_usb/usb_api_register.c
@@ -27,7 +27,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include <hackrf_core.h>
+#include <drivers.h>
 #include <leds.h>
 #include <max283x.h>
 #include <platform_detect.h>
