+++ firmware/hackrf_usb/usb_api_spiflash.c
@@ -26,7 +26,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include <hackrf_core.h>
+#include <drivers.h>
 #include <spi_bus.h>
 #include <usb_queue.h>
 #include <usb_request.h>
