+++ firmware/hackrf_usb/usb_api_sweep.c
@@ -25,8 +25,8 @@
 
 #include <libopencm3/cm3/nvic.h>
 
+#include <drivers.h>
 #include <fixed_point.h>
-#include <hackrf_core.h>
 #include <hackrf_ui.h>
 #include <m0_state.h>
 #include <platform_detect.h>
