+++ firmware/hackrf_usb/usb_api_transceiver.c
@@ -27,8 +27,8 @@
 #include <stddef.h>
 
 #include <clock_gen.h>
+#include <drivers.h>
 #include <fixed_point.h>
-#include <hackrf_core.h>
 #include <leds.h>
 #include <m0_state.h>
 #include <operacake_sctimer.h>
